Lv Exp config more game design battle code prepare

+++ Project/Assets/Scripts/GameDatas/StaticData.cs
@@ -39,7 +39,7 @@ public class MonsterBase
 	public int atk { get; set; }
 	public int def { get; set; }
 	public double atk_spd { get; set; }
-	public int mov_spd { get; set; }
+	public double mov_spd { get; set; }
 	public int range { get; set; }
 	public int intel { get; set; }
 	public int agi { get; set; }
